Updates available

+++ src/main/java/com/example/spring_demo/respositories/CourseRegistrationRepository.java
@@ -8,6 +8,6 @@ import org.springframework.stereotype.Repository;
 @Repository
 @Transactional
 
-public interface CourseRegistrationRepository extends JpaRepository<CourseRegistration, Integer> {
+public interface CourseRegistrationRepository extends JpaRepository<CourseRegistration, Long> {
 
 }
